development for OUT parameter
Modify the create and call behavior of the function, if the function has
the out parameter.

If the function has the out parameter, the return
datatype of the function can be any datatype, and the function return
datatype has no relationship with out pararmeter datatype.

If the function return datatype is not VOID, must specify the RETURN
statement in the function body.

When calling a function in plisql, must specify a variable to receive
the out parameter value.

If turn the guc ivorysql.allow_out_parameter_const off, call a function
in SQL will report an error if the constant or NULL is passed for the out
parameter.
If turn the guc ivorysql.allow_out_parameter_const on, call a function a
SQL, you can pass constant or NULL for the out parameter.

Add a new grammer for anonymous_block,
DO [LANGUAGE lang_name] code [USING IN|OUT|INOUT|IN OUT, ...]

Libpq supports binding parameters by position and name when calling a
function. Some interface functions similar to OCI are implemented.