development for OUT parameter
Modify the create and call behavior of the function, if the function has
the out parameter.

If the function has the out parameter, the returned datatype of the function
can be any datatype, and the function returned datatype has no relationship
with out pararmeter datatype.

If the function returned datatype is not VOID, must provide the RETURN
statement in the function body.

When calling a function in plisql, must specify a variable to carry the
out parameter value.

If turn off the guc ivorysql.allow_out_parameter_const, calling a function
in SQL will report an error if the constant or NULL is passed for the out
parameter.
If turn on the guc ivorysql.allow_out_parameter_const, calling a function a
SQL, passing constant or NULL for the out parameter is allowed.

Add a new grammer for anonymous block:
DO [LANGUAGE lang_name] code [USING IN|OUT|INOUT|IN OUT, ...]

Libpq now supports binding parameters by position or name when calling a
function. Some interface functions similar to OCI are implemented.