test_ddl_deparse: Rename test create_sequence_1 to create_sequence
This test was the only one named following the convention used for
alternate output files.  This was a little bit confusing when looking at
the diffs of the test, because one would think that the diffs are based
on an uncommon case, as alternate outputs are usually used for uncommon
configuration scenarios.

create_sequence_1 was the only test in the tree using such a name, and
it had no alternate output.

Reviewed-by: Kirill Reshke <reshkekirill@gmail.com>
Discussion: https://postgr.es/m/aKLY6wCa_OInr3kY@paquier.xyz