%TYPE and %ROWTYPE enhancement.
If the declaration of the referenced item changes,
then the declaration of the referencing item changes too.

The referencing item inherits data type, data size, constraints
(unless the referenced item is a column) of the referenced item.

Function and prorcedures argument datatype can be %ROWTYPE,
and funtion returned datatype can be %ROWTYPE.